Fix incorrect type
In commit b56452e466719c6ebb37c95a91cfe79b65b8f104, table columns took
on a different range of valid values. Instead of only being able to
specify columns via a string (the axiom name), you can now also use an
axiom directly, or use an array where the first value is the axiom name
and the second value is an object that will be copyFrom'd to override
settings on the axiom.

This commit updates the type so that it's valid to use the second and
third forms listed above.