Fix bug regarding TabbedDetailView width
If there were a enough tabs to overflow the container, then the whole
detail view would become wide enough to fit all of the tabs in one row.
This had the unfortunate side effect of making the content of the tab
overly wide as well, which means the content was getting cut off when
TabbedDetailView is used in the DAO controller.

This commit fixes that issue by making it so that only the row of tabs
is scrollable. The content doesn't expand to match it.